requirements.txt
asgiref==3.8.1
channels==3.0.5
channels_redis==4.2.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.1.0
mysqlclient==2.2.6
PyJWT==2.10.1
redis==5.2.0
sqlparse==0.5.2
tzdata==2024.2
